change: move cart items from cart json column to cart_items table

+++ database/migrations/2020_01_01_181539_create_order_items_table.php
@@ -27,7 +27,7 @@ class CreateOrderItemsTable extends Migration
             $table->unsignedSmallInteger('qty');
             $table->string('filter_details');
             $table->json('filter');
-            $table->unsignedSmallInteger('price');
+            $table->unsignedInteger('price');
             $table->timestamps();
         });
